requirements.txt
flask
pytesseract
pillow
uuid